Now targets .NET 10
Breaking changes due to https://github.com/dotnet/runtime/pull/109690 meaning MailAddress now parsed emails differently. ZeroRedact has been updated to have .NET 8 also have the same behaviour.
Any email with double dots are now handled as invalid email addresses.

Efforts to keep up with MailAddress shouldn't be supported. As long as the the string/readonlyspan version of email redaction continues to be a subset of MailAddress then that should be good enough. MailAddress is more loose and accepts a wider range of email addresses such as "abc123 !@#". The only attempt to keep up with MailAddress is to ensure that all string/readonlyspan validation passes MailAddress validation. It does not need to be the other way around.